Add more unit-tests
This branch's aim is to improve the unit tests to get to > 80%
test coverage.

The only non-test / mechanical change is in the download logic:
this was changed to print all the errors encountered when failing
to download multiple assets as opposed to just the first one.
This is in `internal/lock.go`.