Added spinner, changed symbols
Added a spinner that "spins" (loops through the chars -,\,|,/,-,\) to give a sense of progress, movement, and productivity as the user waits for resources to download.
The spinner is replaced by a checkmark ✔ when all downloads are complete.
The spinner works by using a ticker to continuously place 0's in the progressCh channel, which keeps the bar goroutine running and printing.

Also changed the ASCII symbols that make up the progress bar to look more pleasing.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>