Added "Grabbing it..." message (Temporary)
Next to the bar, added a message reading "Grabbing it..." with a ticking/looping ellipsis.

However, this kind of clutters up the progress bar line, and the ellipsis combined with the spinner adds a bit too much movement, I think.
So will go back to just the spinner and bar.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>