Added counter showing total download size in bytes
A new goroutine loops through the filteredResources and performs HEAD requests to get each resource's metadata.
It totals up the size of all resources in bytes, and that total is displayed for the user.

As of now, calculation of the current number of bytes successfully downloaded hasn't been implemented (hence the <> in "<>/123456B").

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>