Now shows how many bytes have been downloaded
When a resource finishes downloading, its size is added to the bytes-downloaded total that is on-display.
The user now knows exactly how much data has been downloaded.

Before downloading anything, a for loop goes through each resource, fetches its size, and adds it to a list.
	(This for loop is not executed concurrently in a goroutine because a resource could finish downloading before its size is fetched.
	 This is temporary -- concurrency should be restored to improve efficiency.)
When a resource finishes downloading, it looks up its size in the list and places the size in a channel that the print goroutine is monitoring.
The print goroutine adds the incoming sizes to the downloaded total.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>