dont download if file already present
if a file is already i the file system and it validates, we do not need to download it again.

Update resource.go

added comment back to the code

Update resource.go

removed os.Remove(fileName) which was previously on line 82

Update resource.go

deleted my code under the geturltodir function and added the missing line that was requested

dont download if file already present

if a file is already i the file system and it validates, we do not need to download it again.

fixed the too less calls in lock.Download go included a missing boolean value for the lock.Download call.

dont download if file already present

if a file is already i the file system and it validates, we do not need to download it again.

Update resource.go

added comment back to the code

Update resource.go

removed os.Remove(fileName) which was previously on line 82

Update resource.go

deleted my code under the geturltodir function and added the missing line that was requested

dont download if file already present

if a file is already i the file system and it validates, we do not need to download it again.

fixed the too less calls in lock.Download go included a missing boolean value for the lock.Download call.